Fix status code

+++ Dustman/src/main/java/com/dustman/service/UserService.java
@@ -1,7 +1,8 @@
 package com.dustman.service;
 
 import com.dustman.dao.UserDAO;
-import com.dustman.dto.user.*;
+import com.dustman.dto.Status;
+import com.dustman.dto.UserDTO;
 import com.dustman.model.UserDetails;
 import org.springframework.stereotype.Service;
 
@@ -16,50 +17,70 @@ public class UserService {
         this.userDAO = userDAO;
     }
 
-    public String registerUser(RegisterInfo registerInfo) {
+    public Status registerUser(UserDTO registerInfo) {
         String result = userDAO.registerUser(registerInfo);
         if (result.equals("exist")) {
-            return "User already exist";
+            return new Status(409, "User already exist");
         } else if (result.equals("true")) {
-            return "User successfully registered";
+            return new Status(201, "User successfully registered");
         } else {
-            return "Something went wrong! \nPlease try again";
+            return new Status(500, "Internal server error");
         }
     }
 
-    public List<UserDetails> getAllUser() {
-        return userDAO.getAllUser();
-    }
-
-    public UserInfo getUserById(String id) {
-        UserDetails userDetails = userDAO.getUserById(id);
-
-        return new UserInfo(userDetails.getUserId(), userDetails.getUserName(), "Hide it for Security purpose", userDetails.getEmail(), userDetails.getImageUrl(), userDetails.getImageId(), userDetails.getRole(), userDetails.getUserRoc(), userDetails.getCreatedAt(), userDetails.getUpdatedAt());
-    }
-
-    public boolean checkUser(String id) {
-        return userDAO.checkUser(id);
-
-    }
-
-    public String updateUser(UpdatedUserInfo updatedUserInfo) {
-        boolean status = userDAO.updateUser(updatedUserInfo);
-        return (status) ? "User successfully updated" : "Something went wrong! \nPlease try again";
-    }
-
-    public String updatePassword(UpdatePasswordInfo updatePasswordInfo) {
-        if (!userDAO.checkUser(updatePasswordInfo.getUserId())) {
-            return "User not found";
+    public Status getAllUser() {
+        List<UserDetails> userDetails = userDAO.getAllUser();
+        if (!userDetails.isEmpty()) {
+            return new Status(200, userDetails);
         }
-        if (userDAO.checkPassword(updatePasswordInfo.getUserId(), updatePasswordInfo.getOldPassword())) {
-            userDAO.updatePassword(updatePasswordInfo.getUserId(), updatePasswordInfo.getNewPassword());
-            return "Password updated";
+        return new Status(204, "No user found");
+    }
+
+    //todo add Status
+    public Status getUserById(String id) {
+        boolean isExist = userDAO.checkUser(id);
+        if (isExist) {
+            UserDetails userDetails = userDAO.getUserById(id);
+            return new Status(200, userDetails);
+        }
+        return new Status(404, "use not found with this ID");
+    }
+
+    public Status checkUser(String id) {
+        boolean state = userDAO.checkUser(id);
+        if (state) {
+            return new Status(200, "User found");
         } else {
-            return "Old password is incorrect";
+            return new Status(404, "Data not found");
         }
     }
 
-    public String updateRole(UpdateRole updateRole) {
+    public Status updateUser(UserDTO updatedUserInfo) {
+        boolean isExist = userDAO.checkUser(updatedUserInfo.getUserId());
+        if (isExist) {
+            boolean status = userDAO.updateUser(updatedUserInfo);
+            if (status) {
+                return new Status(200, "User successfully updated");
+            }
+            return new Status(500);
+        }
+//
+        return new Status(404, "Data not found for this user ID");
+    }
+
+    public Status updatePassword(UserDTO updatePasswordInfo) {
+        if (userDAO.checkUser(updatePasswordInfo.getUserId())) {
+            if (userDAO.checkPassword(updatePasswordInfo.getUserId(), updatePasswordInfo.getOldPassword())) {
+                userDAO.updatePassword(updatePasswordInfo.getUserId(), updatePasswordInfo.getNewPassword());
+                return new Status(200,"Password Update Successfully");
+            } else {
+                return new Status(400,"Old Password is incorrect");
+            }
+        }
+        return new Status(404, "Data not found for this user ID");
+    }
+
+    public String updateRole(UserDTO updateRole) {
         try {
             if (!userDAO.checkUser(updateRole.getUserId())) {
                 return "User not found";
@@ -75,12 +96,15 @@ public class UserService {
         }
     }
 
-    public String deleteUser(String id) {
-        if (userDAO.deleteUser(id)) {
-            return "User deleted successfully";
-        } else {
-            return "User not found";
+    public Status deleteUser(String id) {
+        if (userDAO.checkUser(id)){
+            if (userDAO.deleteUser(id)) {
+                return new Status(200,"User deleted successfully");
+            } else {
+                return new Status(500);
+            }
         }
+        return new Status(404,"Data not found for this user ID");
     }
 
 }
